backend/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "schedule-concierge"
version = "0.1.0"
description = "Schedule Concierge Backend"
authors = [ { name="Team" } ]
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn>=0.29.0",
  "pydantic>=2.6.0",
  "sqlalchemy>=2.0.28",
  "psycopg2-binary>=2.9.9",
  "alembic>=1.13.1",
  "prometheus-client>=0.20.0",
  "python-dateutil>=2.9.0.post0",
  "passlib[bcrypt]>=1.7.4",
  "python-jose>=3.3.0",
  "python-multipart>=0.0.9",
  "google-auth>=2.28.0",
  "google-auth-oauthlib>=1.2.0",
  "google-api-python-client>=2.120.0",
  "requests>=2.31.0",
  "cryptography>=42.0.0",
  "redis>=5.0.0"
]

[tool.setuptools]
packages = ["app"]

[project.optional-dependencies]
dev = [
  "pytest>=8.1.0",
  "pytest-asyncio>=0.23.5",
  "httpx>=0.27.0",
  "coverage>=7.4.0",
  "opentelemetry-sdk>=1.25.0",
  "opentelemetry-exporter-otlp>=1.25.0",
  "python-dotenv>=1.0.1"
]

[tool.pytest.ini_options]
pythonpath = ["app"]
addopts = "-q --maxfail=1"
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
source = ["app"]

[tool.coverage.report]
fail_under = 80
show_missing = true
